QuickSort
This code quickSorts the reads of a fastq file into alphabetical order. The reads in this file are 44bp long two changes must be made for other fq files. change line 53 for the organism and line 64 for the read length. This can handle a file up to the size of 100M reads efficiently.